@@ -5,7 +5,7 @@ version: 3.0.9+73
 publish_to: none
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.10.0 <3.0.0'
 dependencies:
   flutter:
     sdk: flutter
@@ -14,86 +14,86 @@ dependencies:
 
   # API
   http: ^0.13.4
-  oauth2_client: ^2.3.1
+  oauth2_client: ^2.4.0
   splitwise_api: ^2.0.3
   flutter_google_places: ^0.3.0
-  geocoding: ^2.0.1
-  geolocator: ^8.0.1
+  geocoding: ^2.0.4
+  geolocator: ^8.2.1
 
   # Auth
-  google_sign_in: ^5.2.1
-  sign_in_with_apple: ^3.3.0
+  google_sign_in: ^5.3.1
+  sign_in_with_apple: ^4.0.0
 
   # Data
   faker: ^2.0.0
   flutter_dotenv: ^5.0.2
-  grouped_list: ^4.1.0
+  grouped_list: ^4.2.0
   multiple_stream_builder: ^2.0.0
-  shared_preferences: ^2.0.5
-  sticky_grouped_list: ^2.0.0-nullsafety.2
+  shared_preferences: ^2.0.15
+  sticky_grouped_list: ^2.1.0
 
   # Features
-  add_2_calendar: ^2.1.2
+  add_2_calendar: ^2.1.3
   clipboard: ^0.1.3
-  in_app_review: ^2.0.3
+  in_app_review: ^2.0.4
   maps_launcher: ^2.0.1
-  showcaseview: ^1.1.4
+  showcaseview: ^1.1.5
 
   # Firebase
-  cloud_firestore: ^2.5.4
-  cloud_functions: ^3.1.0
-  firebase_analytics: ^8.3.4
-  firebase_auth: ^3.1.5
-  firebase_core: ^1.9.0
-  firebase_crashlytics: ^2.2.5
-  firebase_messaging: ^11.0.0
-  firebase_performance: ^0.7.1+4
-  firebase_storage: ^10.0.7
+  cloud_firestore: ^3.1.15
+  cloud_functions: ^3.2.15
+  firebase_analytics: ^9.1.8
+  firebase_auth: ^3.3.18
+  firebase_core: ^1.17.0
+  firebase_crashlytics: ^2.8.0
+  firebase_messaging: ^11.4.0
+  firebase_performance: ^0.8.0+12
+  firebase_storage: ^10.2.16
 
   # Images
-  image_cropper: ^1.4.1
-  image_picker: ^0.8.4+4
+  image_cropper: ^2.0.3
+  image_picker: ^0.8.5+3
 
   # Notifications
-  flutter_app_badger: ^1.3.0
-  flutter_local_notifications: ^9.1.5
-  fluttertoast: ^8.0.8
+  flutter_app_badger: ^1.4.0
+  flutter_local_notifications: ^9.5.3+1
+  fluttertoast: ^8.0.9
   notification_permissions: ^0.6.1
 
   # State Management
-  bloc: ^7.0.0
+  bloc: ^8.0.3
   equatable: ^2.0.3
-  flutter_bloc: ^7.2.0
+  flutter_bloc: ^8.0.1
   get_it: ^7.2.0
   observable: ^0.24.0
-  provider: ^6.0.1
+  provider: ^6.0.2
 
   # Self Packaged
   flappy_search_bar:
     path: packages/flappy_search_bar
 
   # Web
-  any_link_preview: ^2.0.4
+  any_link_preview: ^2.0.8
   flutter_linkify: ^5.0.2
-  url_launcher: ^6.0.17
-  webview_flutter: ^3.0.0
+  url_launcher: ^6.1.2
+  webview_flutter: ^3.0.4
 
   # UI
-  animated_splash_screen: ^1.1.0
+  animated_splash_screen: ^1.2.0
   animations: ^2.0.2
   auto_size_text: ^3.0.0
   cupertino_icons: ^1.0.4
-  curved_navigation_bar: ^1.0.1
+  curved_navigation_bar: ^1.0.3
   draggable_scrollbar: ^0.1.0
-  eva_icons_flutter: ^3.0.2
+  eva_icons_flutter: ^3.1.0
   expandable: ^5.0.1
-  flutter_animation_progress_bar: ^2.0.0
+  flutter_animation_progress_bar: ^2.1.1
   flutter_spinkit: ^5.1.0
-  modal_bottom_sheet: ^2.0.0
-  responsive_framework: ^0.1.5
+  modal_bottom_sheet: ^2.0.1
+  responsive_framework: ^0.1.9
   rxdart: ^0.26.0
   sizer: ^2.0.15
-  syncfusion_flutter_gauges: ^19.3.59
+  syncfusion_flutter_gauges: ^20.1.56
 
 
   # Other
