@@ -26,14 +26,14 @@ import '../main_tab_page/main_tab_page.dart';
 class ProfileStream extends StatefulWidget {
   final String uid;
 
-  const ProfileStream({Key key, this.uid,}) : super(key: key);
+  const ProfileStream({Key? key, required this.uid,}) : super(key: key);
 
   @override
   _ProfileStreamState createState() => _ProfileStreamState();
 }
 
 class _ProfileStreamState extends State<ProfileStream> {
-  NotificationBloc bloc;
+  late NotificationBloc bloc;
   final currentUserProfile = locator<UserProfileService>().currentUserProfileDirect();
 
   @override
@@ -64,7 +64,7 @@ class _ProfileStreamState extends State<ProfileStream> {
                 FlutterAppBadger.updateBadgeCount(state.data.length);
                 return MainTabPage(notifications: state.data,);
               } else {
-                return MainTabPage(notifications: null,);
+                return MainTabPage(notifications: [],);
               }}
         ));
   }
