@@ -1,18 +1,18 @@
 class LoginState {
-  final bool? isEmailValid;
-  final bool? isPasswordValid;
-  final bool? isSubmitting;
-  final bool? isSuccess;
-  final bool? isFailure;
+  final bool isEmailValid;
+  final bool isPasswordValid;
+  final bool isSubmitting;
+  final bool isSuccess;
+  final bool isFailure;
 
   bool get isFormValid => isEmailValid! && isPasswordValid!;
 
   LoginState(
-      {this.isEmailValid,
-        this.isPasswordValid,
-        this.isSubmitting,
-        this.isSuccess,
-        this.isFailure});
+      {required this.isEmailValid,
+      required  this.isPasswordValid,
+      required  this.isSubmitting,
+      required  this.isSuccess,
+      required  this.isFailure});
 
   factory LoginState.initial() {
     return LoginState(
