@@ -24,13 +24,13 @@ class PublicProfileRepository {
     UserPublicProfile _userProfileFromSnapshot(DocumentSnapshot snapshot){
       if(snapshot.exists) {
         try {
-          Map<String, dynamic> data = snapshot.data();
+          Map<String, dynamic> data = snapshot.data() as Map<String, dynamic>;
           return UserPublicProfile.fromData(data);
         } catch(e){
           CloudFunction().logError('Error retrieving single user profile:  ${e.toString()}');
-          return null;
+          return UserPublicProfile();
         }} else {
-        return null;
+        return UserPublicProfile();
       }
     }
 
